@@ -19,88 +19,111 @@ namespace Flurl.Test.Http
 		protected abstract IFlurlRequest GetRequest();
 
 		[Test]
+		public async Task can_set_http_version() {
+			Assert.AreEqual("1.1", GetSettings().HttpVersion); // default
+
+			using var test = new HttpTest();
+
+			GetSettings().HttpVersion = "2.0";
+			var req = GetRequest();
+			Assert.AreEqual("2.0", req.Settings.HttpVersion);
+
+			Version versionUsed = null;
+			await req
+				.BeforeCall(c => versionUsed = c.HttpRequestMessage.Version)
+				.GetAsync();
+
+			Assert.AreEqual("2.0", versionUsed.ToString());
+		}
+
+		[Test]
+		public void cannot_set_invalid_http_version() {
+			Assert.Throws<ArgumentException>(() => GetSettings().HttpVersion = "foo");
+		}
+
+		[Test]
 		public async Task can_allow_non_success_status() {
-			using (var test = new HttpTest()) {
-				GetSettings().AllowedHttpStatusRange = "4xx";
-				test.RespondWith("I'm a teapot", 418);
-				try {
-					var result = await GetRequest().GetAsync();
-					Assert.AreEqual(418, result.StatusCode);
-				}
-				catch (Exception) {
-					Assert.Fail("Exception should not have been thrown.");
-				}
+			using var test = new HttpTest();
+
+			GetSettings().AllowedHttpStatusRange = "4xx";
+			test.RespondWith("I'm a teapot", 418);
+			try {
+				var result = await GetRequest().GetAsync();
+				Assert.AreEqual(418, result.StatusCode);
+			}
+			catch (Exception) {
+				Assert.Fail("Exception should not have been thrown.");
 			}
 		}
 
 		[Test]
 		public async Task can_set_pre_callback() {
 			var callbackCalled = false;
-			using (var test = new HttpTest()) {
-				test.RespondWith("ok");
-				GetSettings().BeforeCall = call => {
-					Assert.Null(call.Response); // verifies that callback is running before HTTP call is made
-					callbackCalled = true;
-				};
-				Assert.IsFalse(callbackCalled);
-				await GetRequest().GetAsync();
-				Assert.IsTrue(callbackCalled);
-			}
+			using var test = new HttpTest();
+
+			test.RespondWith("ok");
+			GetSettings().BeforeCall = call => {
+				Assert.Null(call.Response); // verifies that callback is running before HTTP call is made
+				callbackCalled = true;
+			};
+			Assert.IsFalse(callbackCalled);
+			await GetRequest().GetAsync();
+			Assert.IsTrue(callbackCalled);
 		}
 
 		[Test]
 		public async Task can_set_post_callback() {
 			var callbackCalled = false;
-			using (var test = new HttpTest()) {
-				test.RespondWith("ok");
-				GetSettings().AfterCall = call => {
-					Assert.NotNull(call.Response); // verifies that callback is running after HTTP call is made
-					callbackCalled = true;
-				};
-				Assert.IsFalse(callbackCalled);
-				await GetRequest().GetAsync();
-				Assert.IsTrue(callbackCalled);
-			}
+			using var test = new HttpTest();
+
+			test.RespondWith("ok");
+			GetSettings().AfterCall = call => {
+				Assert.NotNull(call.Response); // verifies that callback is running after HTTP call is made
+				callbackCalled = true;
+			};
+			Assert.IsFalse(callbackCalled);
+			await GetRequest().GetAsync();
+			Assert.IsTrue(callbackCalled);
 		}
 
 		[TestCase(true)]
 		[TestCase(false)]
 		public async Task can_set_error_callback(bool markExceptionHandled) {
 			var callbackCalled = false;
-			using (var test = new HttpTest()) {
-				test.RespondWith("server error", 500);
-				GetSettings().OnError = call => {
-					Assert.NotNull(call.Response); // verifies that callback is running after HTTP call is made
-					callbackCalled = true;
-					call.ExceptionHandled = markExceptionHandled;
-				};
-				Assert.IsFalse(callbackCalled);
-				try {
-					await GetRequest().GetAsync();
-					Assert.IsTrue(callbackCalled, "OnError was never called");
-					Assert.IsTrue(markExceptionHandled, "ExceptionHandled was marked false in callback, but exception was not propagated.");
-				}
-				catch (FlurlHttpException) {
-					Assert.IsTrue(callbackCalled, "OnError was never called");
-					Assert.IsFalse(markExceptionHandled, "ExceptionHandled was marked true in callback, but exception was propagated.");
-				}
+			using var test = new HttpTest();
+
+			test.RespondWith("server error", 500);
+			GetSettings().OnError = call => {
+				Assert.NotNull(call.Response); // verifies that callback is running after HTTP call is made
+				callbackCalled = true;
+				call.ExceptionHandled = markExceptionHandled;
+			};
+			Assert.IsFalse(callbackCalled);
+			try {
+				await GetRequest().GetAsync();
+				Assert.IsTrue(callbackCalled, "OnError was never called");
+				Assert.IsTrue(markExceptionHandled, "ExceptionHandled was marked false in callback, but exception was not propagated.");
+			}
+			catch (FlurlHttpException) {
+				Assert.IsTrue(callbackCalled, "OnError was never called");
+				Assert.IsFalse(markExceptionHandled, "ExceptionHandled was marked true in callback, but exception was propagated.");
 			}
 		}
 
 		[Test]
 		public async Task can_disable_exception_behavior() {
-			using (var test = new HttpTest()) {
-				GetSettings().OnError = call => {
-					call.ExceptionHandled = true;
-				};
-				test.RespondWith("server error", 500);
-				try {
-					var result = await GetRequest().GetAsync();
-					Assert.AreEqual(500, result.StatusCode);
-				}
-				catch (FlurlHttpException) {
-					Assert.Fail("Flurl should not have thrown exception.");
-				}
+			using var test = new HttpTest();
+
+			GetSettings().OnError = call => {
+				call.ExceptionHandled = true;
+			};
+			test.RespondWith("server error", 500);
+			try {
+				var result = await GetRequest().GetAsync();
+				Assert.AreEqual(500, result.StatusCode);
+			}
+			catch (FlurlHttpException) {
+				Assert.Fail("Flurl should not have thrown exception.");
 			}
 		}
 
@@ -118,7 +141,7 @@ namespace Flurl.Test.Http
 
 			GetSettings().ResetDefaults();
 
-			Assert.That(GetSettings().JsonSerializer is NewtonsoftJsonSerializer);
+			Assert.That(GetSettings().JsonSerializer is DefaultJsonSerializer);
 			Assert.IsTrue(GetSettings().Redirects.Enabled);
 			Assert.IsNull(GetSettings().BeforeCall);
 			Assert.AreEqual(10, GetSettings().Redirects.MaxAutoRedirects);
@@ -126,17 +149,17 @@ namespace Flurl.Test.Http
 
 		[Test] // #256
 		public async Task explicit_content_type_header_is_not_overridden() {
-			using (var test = new HttpTest()) {
-				// PostJsonAsync will normally set Content-Type to application/json, but it shouldn't touch it if it was set explicitly.
-				await "https://api.com"
-					.WithHeader("content-type", "application/json-patch+json; utf-8")
-					.WithHeader("CONTENT-LENGTH", 10) // header names are case-insensitive
-					.PostJsonAsync(new { foo = "bar" });
+			using var test = new HttpTest();
 
-				var h = test.CallLog[0].HttpRequestMessage.Content.Headers;
-				Assert.AreEqual(new[] { "application/json-patch+json; utf-8" }, h.GetValues("Content-Type"));
-				Assert.AreEqual(new[] { "10" }, h.GetValues("Content-Length"));
-			}
+			// PostJsonAsync will normally set Content-Type to application/json, but it shouldn't touch it if it was set explicitly.
+			await "https://api.com"
+				.WithHeader("content-type", "application/json-patch+json; utf-8")
+				.WithHeader("CONTENT-LENGTH", 10) // header names are case-insensitive
+				.PostJsonAsync(new { foo = "bar" });
+
+			var h = test.CallLog[0].HttpRequestMessage.Content.Headers;
+			Assert.AreEqual(new[] { "application/json-patch+json; utf-8" }, h.GetValues("Content-Type"));
+			Assert.AreEqual(new[] { "10" }, h.GetValues("Content-Length"));
 		}
 	}
 
@@ -170,7 +193,7 @@ namespace Flurl.Test.Http
 			var client2 = new FlurlClient();
 			client2.Settings.Redirects.Enabled = false;
 
-			req.WithClient(client2);
+			req.Client = client2;
 			Assert.IsFalse(req.Settings.Redirects.Enabled, "request should inherit client settings when not set at request level");
 			Assert.AreEqual("4xx", req.Settings.AllowedHttpStatusRange, "request should inherit global settings when not set at request or client level");
 			Assert.AreEqual(123, req.Settings.Redirects.MaxAutoRedirects, "request should inherit global settings when not set at request or client level");
@@ -201,10 +224,15 @@ namespace Flurl.Test.Http
 		}
 
 		[Test]
-		public void can_provide_custom_client_factory() {
-			FlurlHttp.GlobalSettings.HttpClientFactory = new SomeCustomHttpClientFactory();
-			Assert.IsInstanceOf<SomeCustomHttpClient>(GetRequest().Client.HttpClient);
-			Assert.IsInstanceOf<SomeCustomMessageHandler>(GetRequest().Client.HttpMessageHandler);
+		public async Task can_provide_custom_client_factory() {
+			FlurlHttp.GlobalSettings.FlurlClientFactory = new MyCustomClientFactory();
+			var req = GetRequest();
+
+			// client not assigned until request is sent
+			using var test = new HttpTest();
+			await req.GetAsync();
+
+			Assert.IsInstanceOf<MyCustomHttpClient>(req.Client.HttpClient);
 		}
 
 		[Test]
@@ -214,10 +242,17 @@ namespace Flurl.Test.Http
 		}
 
 		[Test]
-		public void can_configure_client_from_FlurlHttp_object() {
+		public async Task can_configure_client_from_FlurlHttp_object() {
 			FlurlHttp.ConfigureClient("http://host1.com/foo", cli => cli.Settings.Redirects.Enabled = false);
-			Assert.IsFalse(new FlurlRequest("http://host1.com/bar").Client.Settings.Redirects.Enabled); // different URL but same host, so should use same client
-			Assert.IsTrue(new FlurlRequest("http://host2.com").Client.Settings.Redirects.Enabled);
+			var req1 = new FlurlRequest("http://host1.com/bar"); // different URL but same host, should use above client
+			var req2 = new FlurlRequest("http://host2.com"); // different host, should use new client
+
+			// client not assigned until request is sent
+			using var test = new HttpTest();
+			await Task.WhenAll(req1.GetAsync(), req2.GetAsync());
+
+			Assert.IsFalse(req1.Client.Settings.Redirects.Enabled);
+			Assert.IsTrue(req2.Client.Settings.Redirects.Enabled);
 		}
 	}
 
@@ -240,24 +275,24 @@ namespace Flurl.Test.Http
 			var ser1 = new FakeSerializer();
 			var ser2 = new FakeSerializer();
 
-			using (var test = new HttpTest()) {
-				var cli = new FlurlClient();
-				cli.Settings.JsonSerializer = ser1;
-				Assert.AreSame(ser1, cli.Settings.JsonSerializer);
+			using var test = new HttpTest();
 
-				var req = new FlurlRequest { Client = cli };
-				Assert.AreSame(ser1, req.Settings.JsonSerializer);
+			var cli = new FlurlClient();
+			cli.Settings.JsonSerializer = ser1;
+			Assert.AreSame(ser1, cli.Settings.JsonSerializer);
 
-				req.Settings.JsonSerializer = ser2;
-				Assert.AreSame(ser2, req.Settings.JsonSerializer);
-			}
+			var req = new FlurlRequest { Client = cli };
+			Assert.AreSame(ser1, req.Settings.JsonSerializer);
+
+			req.Settings.JsonSerializer = ser2;
+			Assert.AreSame(ser2, req.Settings.JsonSerializer);
 		}
 
 		private class FakeSerializer : ISerializer
 		{
 			public string Serialize(object obj) => "foo";
-			public T Deserialize<T>(string s) => default(T);
-			public T Deserialize<T>(Stream stream) => default(T);
+			public T Deserialize<T>(string s) => default;
+			public T Deserialize<T>(Stream stream) => default;
 		}
 	}
 
@@ -268,28 +303,6 @@ namespace Flurl.Test.Http
 
 		protected override FlurlHttpSettings GetSettings() => _client.Value.Settings;
 		protected override IFlurlRequest GetRequest() => _client.Value.Request("http://api.com");
-
-		[Test]
-		public void can_provide_custom_client_factory() {
-			var cli = new FlurlClient();
-			cli.Settings.HttpClientFactory = new SomeCustomHttpClientFactory();
-			Assert.IsInstanceOf<SomeCustomHttpClient>(cli.HttpClient);
-			Assert.IsInstanceOf<SomeCustomMessageHandler>(cli.HttpMessageHandler);
-		}
-
-		[Test]
-		public async Task connection_lease_timeout_creates_new_HttpClient() {
-			var cli = new FlurlClient("http://api.com");
-			cli.Settings.ConnectionLeaseTimeout = TimeSpan.FromMilliseconds(50);
-			var hc = cli.HttpClient;
-
-			await Task.Delay(25);
-			Assert.That(hc == cli.HttpClient);
-
-			// exceed the timeout
-			await Task.Delay(25);
-			Assert.That(hc != cli.HttpClient);
-		}
 	}
 
 	[TestFixture, Parallelizable]
@@ -300,35 +313,6 @@ namespace Flurl.Test.Http
 		protected override FlurlHttpSettings GetSettings() => _req.Value.Settings;
 		protected override IFlurlRequest GetRequest() => _req.Value;
 
-		[Test, NonParallelizable] // #239
-		public void request_default_settings_change_when_client_changes() {
-			FlurlHttp.ConfigureClient("http://test.com", cli => cli.Settings.Redirects.Enabled = false);
-			var req = new FlurlRequest("http://test.com");
-			var cli1 = req.Client;
-			Assert.IsFalse(req.Settings.Redirects.Enabled, "pre-configured client should provide defaults to new request");
-
-			req.Url = "http://test.com/foo";
-			Assert.AreSame(cli1, req.Client, "new URL with same host should hold onto same client");
-			Assert.IsFalse(req.Settings.Redirects.Enabled);
-
-			req.Url = "http://test2.com";
-			Assert.AreNotSame(cli1, req.Client, "new host should trigger new client");
-			Assert.IsTrue(req.Settings.Redirects.Enabled);
-
-			FlurlHttp.ConfigureClient("http://test2.com", cli => cli.Settings.Redirects.Enabled = false);
-			Assert.IsFalse(req.Settings.Redirects.Enabled, "changing client settings should be reflected in request");
-
-			req.Settings = new FlurlHttpSettings();
-			Assert.IsFalse(req.Settings.Redirects.Enabled, "entirely new settings object should still inherit current client settings");
-
-			req.Client = new FlurlClient();
-			Assert.IsTrue(req.Settings.Redirects.Enabled, "entirely new client should provide new defaults");
-
-			req.Url = "http://test.com";
-			Assert.AreNotSame(cli1, req.Client, "client was explicitly set on request, so it shouldn't change even if the URL changes");
-			Assert.IsTrue(req.Settings.Redirects.Enabled);
-		}
-
 		[Test]
 		public void request_gets_global_settings_when_no_client() {
 			var req = new FlurlRequest();
@@ -338,12 +322,10 @@ namespace Flurl.Test.Http
 		}
 	}
 
-	public class SomeCustomHttpClientFactory : IHttpClientFactory
+	class MyCustomClientFactory : DefaultFlurlClientFactory
 	{
-		public HttpClient CreateHttpClient(HttpMessageHandler handler) => new SomeCustomHttpClient();
-		public HttpMessageHandler CreateMessageHandler() => new SomeCustomMessageHandler();
+		public override HttpClient CreateHttpClient(HttpMessageHandler handler) => new MyCustomHttpClient();
 	}
 
-	public class SomeCustomHttpClient : HttpClient { }
-	public class SomeCustomMessageHandler : HttpClientHandler { }
+	class MyCustomHttpClient : HttpClient { }
 }
