@@ -24,6 +24,7 @@ namespace Flurl.Http.Configuration
 			Redirects = new RedirectSettings(this);
 			ResetDefaults();
 		}
+
 		/// <summary>
 		/// Gets or sets the default values to fall back on when values are not explicitly set on this instance.
 		/// </summary>
@@ -41,7 +42,15 @@ namespace Flurl.Http.Configuration
 		}
 
 		/// <summary>
-		/// Gets or sets a pattern representing a range of HTTP status codes which (in addtion to 2xx) will NOT result in Flurl.Http throwing an Exception.
+		/// Gets or sets the HTTP version to be used. Default is "1.1".
+		/// </summary>
+		public string HttpVersion {
+			get => Get<Version>()?.ToString();
+			set => Set(Version.TryParse(value, out var v) ? v : throw new ArgumentException("Invalid HTTP version: " + value));
+		}
+
+		/// <summary>
+		/// Gets or sets a pattern representing a range of HTTP status codes which (in addition to 2xx) will NOT result in Flurl.Http throwing an Exception.
 		/// Examples: "3xx", "100,300,600", "100-299,6xx", "*" (allow everything)
 		/// 2xx will never throw regardless of this setting.
 		/// </summary>
@@ -158,7 +167,7 @@ namespace Flurl.Http.Configuration
 				testVals?.ContainsKey(propName) == true ? (T)testVals[propName] :
 				_vals.ContainsKey(propName) ? (T)_vals[propName] :
 				Defaults != null ? (T)Defaults.Get<T>(propName) :
-				default(T);
+				default;
 		}
 
 		/// <summary>
@@ -170,35 +179,9 @@ namespace Flurl.Http.Configuration
 	}
 
 	/// <summary>
-	/// Client-level settings for Flurl.Http
-	/// </summary>
-	public class ClientFlurlHttpSettings : FlurlHttpSettings
-	{
-		/// <summary>
-		/// Specifies the time to keep the underlying HTTP/TCP connection open. When expired, a Connection: close header
-		/// is sent with the next request, which should force a new connection and DSN lookup to occur on the next call.
-		/// Default is null, effectively disabling the behavior.
-		/// </summary>
-		public TimeSpan? ConnectionLeaseTimeout {
-			get => Get<TimeSpan?>();
-			set => Set(value);
-		}
-
-		/// <summary>
-		/// Gets or sets a factory used to create the HttpClient and HttpMessageHandler used for HTTP calls.
-		/// Whenever possible, custom factory implementations should inherit from DefaultHttpClientFactory,
-		/// only override the method(s) needed, call the base method, and modify the result.
-		/// </summary>
-		public IHttpClientFactory HttpClientFactory {
-			get => Get<IHttpClientFactory>();
-			set => Set(value);
-		}
-	}
-
-	/// <summary>
 	/// Global default settings for Flurl.Http
 	/// </summary>
-	public class GlobalFlurlHttpSettings : ClientFlurlHttpSettings
+	public class GlobalFlurlHttpSettings : FlurlHttpSettings
 	{
 		internal GlobalFlurlHttpSettings() {
 			ResetDefaults();
@@ -209,7 +192,7 @@ namespace Flurl.Http.Configuration
 		/// </summary>
 		public override FlurlHttpSettings Defaults {
 			get => null;
-			set => throw new Exception("Global settings cannot be backed by any higher-level defauts.");
+			set => throw new Exception("Global settings cannot be backed by any higher-level defaults.");
 		}
 
 		/// <summary>
@@ -227,29 +210,15 @@ namespace Flurl.Http.Configuration
 		public override void ResetDefaults() {
 			base.ResetDefaults();
 			Timeout = TimeSpan.FromSeconds(100); // same as HttpClient
-			JsonSerializer = new NewtonsoftJsonSerializer(null);
+			HttpVersion = "1.1";
+			JsonSerializer = new DefaultJsonSerializer();
 			UrlEncodedSerializer = new DefaultUrlEncodedSerializer();
 			FlurlClientFactory = new DefaultFlurlClientFactory();
-			HttpClientFactory = new DefaultHttpClientFactory();
 			Redirects.Enabled = true;
 			Redirects.AllowSecureToInsecure = false;
 			Redirects.ForwardHeaders = false;
 			Redirects.ForwardAuthorizationHeader = false;
 			Redirects.MaxAutoRedirects = 10;
-		}
-	}
-
-	/// <summary>
-	/// Settings overrides within the context of an HttpTest
-	/// </summary>
-	public class TestFlurlHttpSettings : ClientFlurlHttpSettings
-	{
-		/// <summary>
-		/// Resets all test settings to their Flurl.Http-defined default values.
-		/// </summary>
-		public override void ResetDefaults() {
-			base.ResetDefaults();
-			HttpClientFactory = new TestHttpClientFactory();
 		}
 	}
 }
