@@ -1,5 +1,5 @@
-﻿using System.Collections.Generic;
-using System.Linq;
+using System;
+using System.Net.Http;
 
 namespace Flurl.CodeGen
 {
@@ -14,19 +14,21 @@ namespace Flurl.CodeGen
 			if (verb == null)
 				AddArg("verb", "HttpMethod", "The HTTP verb used to make the request.");
 
-			if (HasRequestBody)  {
+			if (HasRequestBody) {
 				if (reqBodyType == "Json")
-					AddArg("data", "object", "An object representing the request body, which will be serialized to JSON.");
+					AddArg("body", "object", "An object representing the request body, which will be serialized to JSON.");
+				else if (reqBodyType == "UrlEncoded")
+					AddArg("body", "object", "An object representing the request body, which will be serialized to a URL-encoded string.");
 				else if (reqBodyType == "String")
-					AddArg("data", "string", "Contents of the request body.");
+					AddArg("body", "string", "The request body.");
 				else if (reqBodyType == null)
-					AddArg("content", "HttpContent", "Contents of the request body.", "null");
+					AddArg("content", "HttpContent", "The request body content.", "null");
 				else
-					AddArg("data", "object", "Contents of the request body.");
+					throw new Exception("how did we get here?");
 			}
 
-			AddArg("cancellationToken", "CancellationToken", "The token to monitor for cancellation requests.", "default(CancellationToken)");
-			AddArg("completionOption", "HttpCompletionOption", "The HttpCompletionOption used in the request. Optional.", "HttpCompletionOption.ResponseContentRead");
+			AddArg("completionOption", "HttpCompletionOption", "The HttpCompletionOption used in the request. Optional.", $"HttpCompletionOption.{DefaultHttpCompletionOption}");
+			AddArg("cancellationToken", "CancellationToken", "The token to monitor for cancellation requests.", "default");
 
 			Returns($"Task<{TaskArg}>", $"A Task whose result is {ReturnTypeDescription}.");
 		}
@@ -44,31 +46,25 @@ namespace Flurl.CodeGen
 		public string RequestBodyType { get; }
 		public string ResponseBodyType { get; }
 
-		public string TaskArg {
-			get {
-				switch (ResponseBodyType) {
-					case "Json": return IsGeneric ? "T" : "dynamic";
-					case "JsonList": return "IList<dynamic>";
-					case "String": return "string";
-					case "Stream": return "Stream";
-					case "Bytes": return "byte[]";
-					default: return "IFlurlResponse";
-				}
-			}
-		}
+		public string TaskArg => ResponseBodyType switch {
+			"Json" => "T",
+			"String" => "string",
+			"Stream" => "Stream",
+			"Bytes" => "byte[]",
+			_ => "IFlurlResponse"
+		};
 
-		public string ReturnTypeDescription {
-			get {
-				//var response = (xm.DeserializeToType == null) ? "" : "" + xm.TaskArg;
-				switch (ResponseBodyType) {
-					case "Json": return "the JSON response body deserialized to " + (IsGeneric ? "an object of type T" : "a dynamic");
-					case "JsonList": return "the JSON response body deserialized to a list of dynamics";
-					case "String": return "the response body as a string";
-					case "Stream": return "the response body as a Stream";
-					case "Bytes": return "the response body as a byte array";
-					default: return "the received IFlurlResponse";
-				}
-			}
-		}
+		public string ReturnTypeDescription => ResponseBodyType switch {
+			"Json" => "the JSON response body deserialized to an object of type T",
+			"String" => "the response body as a string",
+			"Stream" => "the response body as a Stream",
+			"Bytes" => "the response body as a byte array",
+			_ => "the received IFlurlResponse"
+		};
+
+		public HttpCompletionOption DefaultHttpCompletionOption => ResponseBodyType switch {
+			"Stream" => HttpCompletionOption.ResponseHeadersRead, // #630
+			_ => HttpCompletionOption.ResponseContentRead
+		};
 	}
 }
