@@ -16,12 +16,13 @@ namespace Flurl.Http
 		/// </summary>
 		/// <param name="buildContent">A delegate for building the content parts.</param>
 		/// <param name="request">The IFlurlRequest.</param>
+		/// <param name="completionOption">The HttpCompletionOption used in the request. Optional.</param>
 		/// <param name="cancellationToken">The token to monitor for cancellation requests.</param>
 		/// <returns>A Task whose result is the received IFlurlResponse.</returns>
-		public static Task<IFlurlResponse> PostMultipartAsync(this IFlurlRequest request, Action<CapturedMultipartContent> buildContent, CancellationToken cancellationToken = default(CancellationToken)) {
+		public static Task<IFlurlResponse> PostMultipartAsync(this IFlurlRequest request, Action<CapturedMultipartContent> buildContent, HttpCompletionOption completionOption = HttpCompletionOption.ResponseContentRead, CancellationToken cancellationToken = default) {
 			var cmc = new CapturedMultipartContent(request.Settings);
 			buildContent(cmc);
-			return request.SendAsync(HttpMethod.Post, cmc, cancellationToken);
+			return request.SendAsync(HttpMethod.Post, cmc, completionOption, cancellationToken);
 		}
 	}
 }
