@@ -1,8 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Net.Http;
-using System.Threading;
 using System.Threading.Tasks;
 
 namespace Flurl.Http
@@ -21,36 +18,9 @@ namespace Flurl.Http
 		/// <example>x = await url.PostAsync(data).ReceiveJson&lt;T&gt;()</example>
 		/// <exception cref="FlurlHttpException">Condition.</exception>
 		public static async Task<T> ReceiveJson<T>(this Task<IFlurlResponse> response) {
-			using (var resp = await response.ConfigureAwait(false)) {
-				if (resp == null) return default(T);
-				return await resp.GetJsonAsync<T>().ConfigureAwait(false);
-			}
-		}
-
-		/// <summary>
-		/// Deserializes JSON-formatted HTTP response body to a dynamic object. Intended to chain off an async call.
-		/// </summary>
-		/// <returns>A Task whose result is a dynamic object containing data in the response body.</returns>
-		/// <example>d = await url.PostAsync(data).ReceiveJson()</example>
-		/// <exception cref="FlurlHttpException">Condition.</exception>
-		public static async Task<dynamic> ReceiveJson(this Task<IFlurlResponse> response) {
-			using (var resp = await response.ConfigureAwait(false)) {
-				if (resp == null) return null;
-				return await resp.GetJsonAsync().ConfigureAwait(false);
-			}
-		}
-
-		/// <summary>
-		/// Deserializes JSON-formatted HTTP response body to a list of dynamic objects. Intended to chain off an async call.
-		/// </summary>
-		/// <returns>A Task whose result is a list of dynamic objects containing data in the response body.</returns>
-		/// <example>d = await url.PostAsync(data).ReceiveJsonList()</example>
-		/// <exception cref="FlurlHttpException">Condition.</exception>
-		public static async Task<IList<dynamic>> ReceiveJsonList(this Task<IFlurlResponse> response) {
-			using (var resp = await response.ConfigureAwait(false)) {
-				if (resp == null) return null;
-				return await resp.GetJsonListAsync().ConfigureAwait(false);
-			}
+			using var resp = await response.ConfigureAwait(false);
+			if (resp == null) return default;
+			return await resp.GetJsonAsync<T>().ConfigureAwait(false);
 		}
 
 		/// <summary>
@@ -59,10 +29,9 @@ namespace Flurl.Http
 		/// <returns>A Task whose result is the response body as a string.</returns>
 		/// <example>s = await url.PostAsync(data).ReceiveString()</example>
 		public static async Task<string> ReceiveString(this Task<IFlurlResponse> response) {
-			using (var resp = await response.ConfigureAwait(false)) {
-				if (resp == null) return null;
-				return await resp.GetStringAsync().ConfigureAwait(false);
-			}
+			using var resp = await response.ConfigureAwait(false);
+			if (resp == null) return null;
+			return await resp.GetStringAsync().ConfigureAwait(false);
 		}
 
 		/// <summary>
@@ -90,10 +59,9 @@ namespace Flurl.Http
 		/// <returns>A Task whose result is the response body as a byte array.</returns>
 		/// <example>bytes = await url.PostAsync(data).ReceiveBytes()</example>
 		public static async Task<byte[]> ReceiveBytes(this Task<IFlurlResponse> response) {
-			using (var resp = await response.ConfigureAwait(false)) {
-				if (resp == null) return null;
-				return await resp.GetBytesAsync().ConfigureAwait(false);
-			}
+			using var resp = await response.ConfigureAwait(false);
+			if (resp == null) return null;
+			return await resp.GetBytesAsync().ConfigureAwait(false);
 		}
 	}
 }
