@@ -17,7 +17,7 @@ namespace Flurl.Http
 		/// <param name="client">The IFlurlClient.</param>
 		/// <param name="action">Action defining the settings changes.</param>
 		/// <returns>The IFlurlClient with the modified Settings</returns>
-		public static IFlurlClient Configure(this IFlurlClient client, Action<ClientFlurlHttpSettings> action) {
+		public static IFlurlClient Configure(this IFlurlClient client, Action<FlurlHttpSettings> action) {
 			action(client.Settings);
 			return client;
 		}
@@ -28,19 +28,8 @@ namespace Flurl.Http
 		/// <param name="request">The IFlurlRequest.</param>
 		/// <param name="action">Action defining the settings changes.</param>
 		/// <returns>The IFlurlRequest with the modified Settings</returns>
-		public static IFlurlRequest ConfigureRequest(this IFlurlRequest request, Action<FlurlHttpSettings> action) {
+		public static IFlurlRequest WithSettings(this IFlurlRequest request, Action<FlurlHttpSettings> action) {
 			action(request.Settings);
-			return request;
-		}
-
-		/// <summary>
-		/// Fluently specify the IFlurlClient to use with this IFlurlRequest.
-		/// </summary>
-		/// <param name="request">The IFlurlRequest.</param>
-		/// <param name="client">The IFlurlClient to use when sending the request.</param>
-		/// <returns>A new IFlurlRequest to use in calling the Url</returns>
-		public static IFlurlRequest WithClient(this IFlurlRequest request, IFlurlClient client) {
-			request.Client = client;
 			return request;
 		}
 
